refactor: extract useQuestions()
